feat: refactor page structure and add tab components for better organization of upload, summary, and flashcard functionalities
