+++ frontend/src/components/tabs/SummaryTab.tsx
@@ -1,0 +1,70 @@
+'use client'
+
+import { Summary } from "@/types/documentTypes";
+
+interface SummaryTabProps {
+    summary: Summary | null;
+    isLoading: boolean;
+    handleReset: () => void;
+}
+
+export default function SummaryTab({ summary, isLoading, handleReset }: SummaryTabProps) {
+    return (
+    <div className="bg-white dark:bg-gray-800 rounded-xl shadow-xl p-8">
+        <div className="flex items-center justify-between mb-6">
+            <h2 className="text-2xl font-semibold text-gray-800 dark:text-white">
+                AI-Generated Summary
+            </h2>
+            <button
+                onClick={handleReset}
+                className="px-4 py-2 text-sm border border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 rounded-lg hover:bg-gray-50 dark:hover:bg-gray-700 transition-colors"
+            >
+                New Notes
+            </button>
+        </div>
+
+        {isLoading ? (
+            <div className="flex items-center justify-center py-12">
+                <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
+                    <span className="ml-4 text-gray-600 dark:text-gray-300">
+                        Loading ...
+                    </span>
+            </div>
+        ) : summary ? (
+            <div className="prose dark:prose-invert max-w-none">
+                <div className="bg-gradient-to-r from-blue-50 to-indigo-50 dark:from-blue-900/20 dark:to-indigo-900/20 p-6 rounded-lg border-l-4 border-blue-500">
+                    <h3 className="text-xl font-semibold mb-4 text-gray-800 dark:text-white">
+                        {summary.title}
+                    </h3>
+                    <ul className="space-y-2">
+                        {summary.keyPoints.map((point, index) => (
+                            <li key={index} className="text-gray-700 dark:text-gray-300">
+                                {point}
+                            </li>
+                        ))}
+                    </ul>
+                </div>
+
+
+                <div className="mt-8 flex gap-4">
+                    <button className="flex-1 px-6 py-3 bg-gradient-to-r from-purple-500 to-pink-600 text-white rounded-lg hover:from-purple-600 hover:to-pink-700 transition-all transform hover:scale-105">
+                        Generate Flashcards
+                    </button>
+                    <button className="flex-1 px-6 py-3 bg-gradient-to-r from-purple-500 to-pink-600 text-white rounded-lg hover:from-purple-600 hover:to-pink-700 transition-all transform hover:scale-105">
+                        Create Quiz
+                    </button>
+                    <button className="flex-1 px-6 py-3 bg-gradient-to-r from-purple-500 to-pink-600 text-white rounded-lg hover:from-purple-600 hover:to-pink-700 transition-all transform hover:scale-105">
+                        Coding Problems
+                    </button>
+                </div>
+            </div>
+        ) : (
+            <div className="text-center py-12">
+                <p className="text-gray-500 dark:text-gray-400">
+                    No summary available. Please upload your notes first.
+                </p>
+            </div>
+        )}
+    </div>
+    );
+}
